# Optimize Minimum VersionVector Computation for Performance