--- a/server/backend/database/memory/database.go
+++ b/server/backend/database/memory/database.go
@@ -1360,29 +1360,20 @@ func (d *DB) UpdateAndFindMinSyncedVersionVector(
 		return nil, fmt.Errorf("find all version vectors: %w", err)
 	}
 
-	// 02. Compute min version vector.
-	var minVersionVector time.VersionVector
-
-	// 02-1. Compute min version vector of other clients and collect attachedActorIDs.
+	var versionVectorInfos []database.VersionVectorInfo
 	for raw := iterator.Next(); raw != nil; raw = iterator.Next() {
 		vvi := raw.(*database.VersionVectorInfo)
-		if clientInfo.ID == vvi.ClientID {
-			continue
-		}
-
-		if minVersionVector == nil {
-			minVersionVector = vvi.VersionVector
-			continue
-		}
-
-		minVersionVector = minVersionVector.Min(vvi.VersionVector)
+		versionVectorInfos = append(versionVectorInfos, *vvi)
 	}
+
+	// 02-1. Compute min version vector of other clients.
+	minVersionVector := database.FindMinVersionVector(versionVectorInfos, clientInfo.ID)
+	// 02-2. Compute min version vector with current client's version vector.
 	if minVersionVector == nil {
 		minVersionVector = versionVector
+	} else {
+		minVersionVector = minVersionVector.Min(versionVector)
 	}
-
-	// 02-2. Compute min version vector with current client's version vector.
-	minVersionVector = minVersionVector.Min(versionVector)
 
 	// 03. Update current client's version vector. If the client is detached, remove it.
 	// This is only for the current client and does not affect the version vector of other clients.