<!--  Thanks for sending a pull request! -->

**What this PR does / why we need it**:

This PR optimizes the minimum version vector (MinVV) computation.
The comparison results of BenchmarkVersionVector test with 1000 clients are as follows:

| Metric | Before | After | Change |
|--------|---------|---------|---------|
| Execution Time | 72.28s | 57.48s | -20.5% |
| Memory Usage | 46.79 GB | 19.85 GB | -57.6% |
| Allocation Count | 108.58M | 94.51M | -13.0% |
| Push/Pull Time | 165 ms | 79 ms | -52.1% |
| Attach Time | 153 ms | 145 ms | -5.2% |
| Post-detach Push/Pull | 38 ms | 33 ms | -13.2% |

- **Memory Optimization**: Reduced memory usage by 57.6% (from 46.79GB to 19.85GB)
- **Push/Pull Performance**: Improved push/pull operation speed by 52.1%
- **Overall Time**: Decreased total execution time by 20.5%

| cpu.prof | mem.prof |
| :--: | :--: | 
| <img width="1297" alt="image" src="https://github.com/user-attachments/assets/7cd8c5d9-02c2-4e89-84ea-21b044f4edf5" />| <img width="1421" alt="image" src="https://github.com/user-attachments/assets/869c48e9-fdbe-4d27-820b-2ec716059597" />

 

**Which issue(s) this PR fixes**:

<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Fixes #1152 

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:

<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->

```release-note

```

**Additional documentation**:

<!--
This section can be blank if this pull request does not require a release note.

Please use the following format for linking documentation:
- [Usage]: <link>
- [Other doc]: <link>
-->

```docs

```

**Checklist**:

- [x] Added relevant tests or not required
- [x] Addressed and resolved all CodeRabbit review comments
- [x] Didn't break anything


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Tests**
  - Expanded test coverage now ensures robust validation of data synchronization and versioning logic across multiple scenarios.
- **Refactor**
  - Streamlined internal mechanisms for computing minimum version vectors to enhance reliability and maintainability, contributing to a more consistent user experience.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->